fix: remove duplicate js_run_devserver.mjs file
